a POST test works (!) and all rows in a database are being saved.

+++ src/db/db.py
@@ -1,8 +1,11 @@
-
 from sqlalchemy.ext.asyncio import create_async_engine
 from sqlalchemy.orm import sessionmaker
 from sqlmodel import SQLModel
 from sqlmodel.ext.asyncio.session import AsyncSession
+
+
+class Init:
+    init = False
 
 
 async def prepare_engine(dsn: str, echo=False):
@@ -14,8 +17,10 @@ async def prepare_engine(dsn: str, echo=False):
 
     # init models
     async with engine.begin() as conn:
-        await conn.run_sync(SQLModel.metadata.drop_all)
-        await conn.run_sync(SQLModel.metadata.create_all)
+        if Init.init is False:
+            await conn.run_sync(SQLModel.metadata.drop_all)
+            await conn.run_sync(SQLModel.metadata.create_all)
+            Init.init = True
 
     return engine
